Frota_Bella/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.12
aiofiles==24.1.0
reportlab>=4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
